% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{bad_wth_check}
\alias{bad_wth_check}
\title{Check for bad weather file}
\usage{
bad_wth_check(.w_fname, .row, .gridid, .ssp, .gcm)
}
\arguments{
\item{.w_fname}{character, weather file}

\item{.row}{integer, row number in simulation data.table}

\item{.gridid}{integer, gridid associated with row in data.table}

\item{.ssp}{character, ssp of row in data.table}

\item{.gcm}{character, gcm of row in data.table}
}
\description{
This function checks if the weather file has less than the required 7 columns
and breaks the simulation loop for gridid associated with this weather file
}
