% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{biov}
\alias{biov}
\title{Calculate bio-climatic variables}
\usage{
biov(prec, tmin, tmax)
}
\arguments{
\item{prec}{vector of daily precipitation}

\item{tmin}{vector of daily minimum temperature}

\item{tmax}{vector of daily maximum temperature}
}
\value{
A named list (bio1... bio19), length  19 of the following bioclimatic variables
 - bio1 : Annual Mean Temperature
 - bio2 : Mean Diurnal Range
 - bio4 : Temperature Seasonality
 - bio5 : Max Temperature of Warmest Period
 - bio6 : Min Temperature of Coldest Period
 - bio7 : Temperature Annual Range (P5-P6)
 - bio3 : Isothermality (P2 / P7)
 - bio8 : Mean Temperature of Wettest Quarter
 - bio9 : Mean Temperature of Driest Quarter
 - bio10: Mean Temperature of Warmest Quarter
 - bio11: Mean Temperature of Coldest Quarter
 - bio12: Annual Precipitation
 - bio13: Precipitation of Wettest Period
 - bio14: Precipitation of Driest Period
 - bio15: Precipitation Seasonality
 - bio16: Precipitation of Wettest Quarter
 - bio17: Precipitation of Driest Quarter
 - bio18: Precipitation of Warmest Quarter
 - bio19: Precipitation of Coldest Quarter
}
\description{
Calculates bioclimatic variables from daily values for Tmin, Tmax, and precipitation
}
