% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{calc_weather_stats}
\alias{calc_weather_stats}
\title{Calculate weather statistics}
\usage{
calc_weather_stats(
  w,
  ssp,
  gcm,
  cell,
  plant_day,
  harvest_day,
  NY = 10,
  prob = c(0.05, 0.1, 0.25, 0.5, 0.75, 0.9, 0.95),
  bins.tmax = c(-Inf, seq(0, 35, by = 5), Inf),
  bins.tmin = c(-Inf, seq(0, 35, by = 5), Inf),
  bins.pr = c(0, 2, 4, 7, 13, 24, 44, 81, 150, Inf)
)
}
\arguments{
\item{w}{data.table of weather file}

\item{NY}{Number of years in block (default = 10)}

\item{prob}{probability bins for quantile decomposition (default = c(0.05, 0.1, 0.25, 0.5, 0.75, 0.9, 0.95))}

\item{bins.tmax}{bin boundaries for histogram decomposition of tmax. Default = c(-Inf, seq(0, 35, by = 5), Inf)}

\item{bins.tmin}{bin boundaries for histogram decomposition of tmin. Default = c(-Inf, seq(0, 35, by = 5), Inf).
Default is based on https://journals-ametsoc-org.proxy.library.cornell.edu/view/journals/clim/35/17/JCLI-D-21-0617.1.xml
by which we use power law breaks to characterize precip, with <=4mm/day as cutoff
using round(10^seq.int(log10(2), log10(150), length.out = 8))}

\item{bins.pr}{bin boundaries for histogram decomposition of precipitation. Default = c(0, 2, 4, 7, 13, 24, 44, 81, 150, Inf)}
}
\description{
Used to calculate summary statistics of weather file data, over successive periods of time (default is decadal).
}
