% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_data_site}
\alias{cell_data_site}
\title{Extended site.100 files.}
\format{
## `cell_data_site`
A data frame with 87387 rows and 7 columns:
\describe{
  \item{gridid}{Grid ID}
  \item{regionid}{Region ID}
  \item{crop}{Crop, four levels}
  \item{irr}{Irrigation, binary}
  \item{run_seq}{Run sequence, identified from land use cohort selection process.}
  \item{eq_schl_name}{Equilibrium schedule file code.}
  \item{site100}{Extended site.100 data.}
  ...
}
}
\source{
Colorado State University, Yi Yang.
}
\usage{
cell_data_site
}
\description{
The extended site.100 files for each crop x irrigation x gridid combination.
Extended site files identified a priori based on land use cohort selection criteria
for simulations.
}
\keyword{datasets}
