% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cmip6_calendars}
\alias{cmip6_calendars}
\title{Soil Futures Datasets}
\format{
## `cmip6_calendars`
A data frame with 39 rows and 2 columns:
\describe{
  \item{gcm}{General Circulation Model}
  \item{calendar}{Calendar type; 5 levels}
  ...
}
}
\source{
NASA, Bridget Thrasher.
}
\usage{
cmip6_calendars
}
\description{
CMIP6 Calendars.
}
\details{
Calendar types of the CMIP6 models.
}
\keyword{datasets}
