% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cover_crop_template}
\alias{cover_crop_template}
\title{Cover Crop Template.}
\format{
## `cover_crop_template`
A data frame with 11982 rows and 1 column:
\describe{
  \item{V1}{crop.100 parameters; start of crop indicated by crop code in all caps}
  ...
}
}
\source{
Colorado State University, Yi Yang.
}
\usage{
cover_crop_template
}
\description{
crop.100 parameters for DayCent simulations. Starts with cover crops and followed by
crops required for rewild scenarios.
}
\keyword{datasets}
