% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covercrop_schedule_template}
\alias{covercrop_schedule_template}
\title{Cover Crop Schedule Template.}
\format{
## `schedule_template`
A data frame with 102828 rows and 8 columns:
\describe{
  \item{scenario}{Scenario in simulation; four levels}
  \item{N_or_S}{North or South Hemisphere - loosely defined.}
  \item{irr}{Irrgiation, binary}
  \item{maiz}{Maize, binary}
  \item{soyb}{Soybean, binary.}
  \item{swht}{Spring wheat, binary.}
  \item{harv_error}{Harvest error in N hemisphere, binary.}
  \item{schedule}{Schedule lines relevant for the scenario.}
  ...
}
}
\source{
Cornell University, S.C. McClelland
}
\usage{
covercrop_schedule_template
}
\description{
Cover crop schedule assignments for scenarios in DayCent simulations. Follows sch file format.
Does not include rewild scenario. Accounts for growing degree day harvest dates.
}
\keyword{datasets}
