% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_spatial.R
\name{create_cell_data_csu_table}
\alias{create_cell_data_csu_table}
\title{Create a data.table containing all variables required to run a DayCent simulation}
\usage{
create_cell_data_csu_table(cmip6_calendars)
}
\description{
Returns a large table of all data required to run global simulations with CSU spin-ups.
}
\details{
This function generates a data.table with rows for every combination of scenario, crop and ssp
and the raster cell numbers corresponding to locations where that crop is grown globally.
Data values are automatically populated using spatial crop calendar, crop management, and soil data.
}
