% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{create_crop}
\alias{create_crop}
\title{Create a crop.100 file}
\usage{
create_crop(cell_data, cell, cover_crop = copy(cover_crop_template), tmp.dir)
}
\arguments{
\item{cell_data}{multiple row data.table providing input data for the simulation.}

\item{cell}{unique cell value to be passed to function.}

\item{cover_crop}{table template of cover crop parameters to include in the crop.100.}

\item{tmp.dir}{directory to create in tmp.path, set with arg[1]}
}
\description{
Used for writing a crop.100 file based on Yang et al. (2022) ERL
}
\details{
This function accepts parameter values from a csv file and creates a crop.100 for
all simulations for a crop type in a cell.
}
