% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site.R
\name{create_eq_sch}
\alias{create_eq_sch}
\title{Create a DayCent equilibrium schedule file}
\usage{
create_eq_sch(
  .gridid,
  .schl_name,
  .schl_seq,
  EQ_sch_data = copy(EQ_sch_data),
  tmp.dir
)
}
\arguments{
\item{.gridid}{integer, grid cell ID number}

\item{.schl_name}{character, name of land use ID associated with grid cell ID number}

\item{.schl_seq}{integer, schedule file number in the equilibrium run sequence}

\item{EQ_sch_data}{data table, contains data associated with each grid cell and its 2 eq sch files}

\item{tmp.dir}{character, temporary directory to write the schedule files to}
}
\description{
Used for extracting a schedule file from a data table.
}
\details{
This function creates a universal equilibrium schedule file for a grid cell based on a land use ID.
}
