% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{create_sched}
\alias{create_sched}
\title{Create a DayCent Schedule file}
\usage{
create_sched(
  cell_data,
  schedule_table = copy(schedule_template),
  ssp,
  gcm,
  crop,
  scenario,
  weather_filename
)
}
\arguments{
\item{cell_data}{single row data.table providing input data for the simulation.
If more than one row is provided, only first will be used}

\item{schedule_table}{data.table providing template of schedule file.}

\item{ssp}{character, specifies scenario}

\item{gcm}{character, specifies scenario}

\item{crop}{character, one of 'maize', 'wwheat', 'swheat', 'soy'}

\item{scenario}{character, specifies scenario}
}
\value{
invisibly returns boolean indicating whether file was written successfully.
}
\description{
Used for side effect of writing a schedule file based on spatial data.
}
\details{
This function creates a schedule file for a specified scenario and crop,
using spatial climate, crop calendar, crop management, and soil data.
}
