% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site.R
\name{create_site}
\alias{create_site}
\title{Create a DayCent site file}
\usage{
create_site(
  .gridid,
  .run_seq,
  .irr,
  .crop,
  cell_data_site = copy(cell_data_site),
  tmp.dir
)
}
\arguments{
\item{.gridid}{integer, grid cell ID number}

\item{.run_seq}{integer, land use cohort ID number}

\item{.irr}{Integer, irrigation level}

\item{.crop}{character, name of crop aligned with gridid and run_seq}

\item{cell_data_site}{Data table, contains spatial data associated with each cell and site.100 file}

\item{tmp.dir}{character, temporary directory to write the site.100 file to}
}
\description{
Used for writing a site file using spatial data.
}
\details{
This function creates a universal site file for a grid cell.
}
