% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eq_pnh_sch_template}
\alias{eq_pnh_sch_template}
\title{Northern Hemisphere Rewild Schedule Files.}
\format{
## `eq_pnh_sch_template`
A data frame with 1971 rows and 2 columns:
\describe{
  \item{schl_name}{Schedule File Name}
  \item{schl_data}{Schedule lines relevant for the land use type.}
  ...
}
}
\source{
Colorado State University, Yi Yang. Cornell University, S.C. McClelland.
}
\usage{
eq_pnh_sch_template
}
\description{
The extended site.100 files for each crop x irrigation x gridid combination.
Extended site files identified a priori based on land use cohort selection criteria
for simulations.
}
\keyword{datasets}
