% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sql.R
\name{establish.con}
\alias{establish.con}
\title{Create a connection to MySQL database}
\usage{
establish.con(db.name, host.name, port.no, user.name)
}
\arguments{
\item{db.name}{is database name (ie 'GlobSim_hist)}

\item{host.name}{is host name (ie 'localhost)}

\item{port.no}{is port number (ie 3306)}

\item{user.name}{is user name ('shelby')}
}
\description{
Create a connection to MySQL database
}
