% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{initialize_weather}
\alias{initialize_weather}
\title{Create a data table with structure for wth file}
\usage{
initialize_weather(start_year, end_year)
}
\arguments{
\item{start_year}{Character, specifying the start year of the simulation.}

\item{end_year}{Character, specifying the end year of the simulation.}
}
\description{
Create a data table with structure for wth file
}
