% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{load_climate}
\alias{load_climate}
\title{Loads climate data}
\usage{
load_climate(ssp, gcm)
}
\arguments{
\item{ssp}{Chareacter secifying the ssp.}

\item{gcm}{Character, specifying the climate model (gcm).}
}
\value{
list of rasters (see details)
}
\description{
For specified gcm and ssp, creates a list of global daily rasters.  Each element of the list corresponds to a specific climate variable.
These are mean daily temperature (tas), minimum daily temperature (tasmin), maximum daily temperature (tasmax),
and precipitation (pr).  Each layer of the rasters corresponds to a day from 2015 to 2100 (using the gcms calendar)
}
