% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{load_climate_25deg}
\alias{load_climate_25deg}
\title{Loads climate data for 0.25 deg}
\usage{
load_climate_25deg(ssp, gcm, start_year, end_year)
}
\arguments{
\item{ssp}{Chareacter secifying the ssp.}

\item{gcm}{Character, specifying the climate model (gcm).}
}
\value{
list of rasters (see details)
}
\description{
For specified gcm and ssp, creates a list of global daily rasters.  Each element of the list corresponds to a specific climate variable.
These are mean daily temperature (tas), minimum daily temperature (tasmin), maximum daily temperature (tasmax),
and precipitation (pr).  Each layer of the rasters corresponds to a day from a specified start year to end year (using the gcms calendar)
}
