% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{make_blocks2}
\alias{make_blocks2}
\title{Rounding to n year increments}
\usage{
make_blocks2(x, n = 5)
}
\arguments{
\item{x, }{number}

\item{n, }{integer, number of years in increment, default is 5}
}
\description{
Rounding to n year increments
}
