% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{make_weather_file_25deg}
\alias{make_weather_file_25deg}
\title{Create a weather file from 0.25 degree data}
\usage{
make_weather_file_25deg(
  climate,
  .gridid,
  .xy,
  .gcm,
  .ssp,
  weather = pkg.env$weather_data,
  cmip6_calendars = cmip6_calendars,
  tmp.dir,
  out.dir
)
}
\arguments{
\item{climate}{list of spatial rast objects.
One element of list for each of the climate variables (tasmin, tasmax, pr).
Usually created using load_climate()}

\item{.gridid}{integer, cell number (-180 to 180 longitude) from cell_data_table.}

\item{.xy}{dataframe, specifying the rotated x coordinate and unrotated y coordinate from which to extract values.
Note that climate rasters are in rotated (0-360 degree longitude) coordinates.}

\item{.gcm}{character (length one) giving name of the cmip6 model.}

\item{.ssp}{character (length one) giving name of ssp scenario.}

\item{weather}{data.table containing gregorian calendar and index numbers to extract gregorian values
from a daily climate rast that uses a different calendar. Usually created using initialize_weather()}

\item{tmp.dir}{directory to create in tmp.path, set with arg[1]}

\item{out.dir}{directory to create in out.path, set with arg[2]}
}
\description{
Used for side effect of writing a weather file based on location.
Location is specified as a raster cell number.
}
\details{
This function creates a weather file for a specified climate model, ssp, and location.
Raw climate model data is converted into daily values in the Gregorian calendar, which sometimes
requires date conversion from the native gcm calendar.  Daily vectors are constructed from 2015 through
to 2100 for each of, minimum daily temperature (tasmin), maximum daily temperature (tasmax),
and precipitation (pr).  These vectors are written as columns in the weather file which is written to
a uniquely named file.  File naming is unique to allow parallel execution of DayCent, with each instance
using a different weather file.
}
