% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{make_weather_grwszn_stats}
\alias{make_weather_grwszn_stats}
\title{Create weather growing and non-growing season statistics file}
\usage{
make_weather_grwszn_stats(
  weather_fname,
  .gridid,
  .crop,
  .ssp,
  .gcm,
  .plant_date,
  .harv_date,
  out.dir
)
}
\arguments{
\item{weather_fname}{data.table, containing 86 years (2016-2100) of tasmin, tasmax, and pr data for gridcell.
Must be read in to function as a file.path}

\item{.gridid}{integer, cell number (-180 to 180 longitude) from cell_data_table.}

\item{.crop}{character (length one) giving crop name from cell_data_table.}

\item{.ssp}{character (length one) giving name of ssp scenario.}

\item{.gcm}{character (length one) giving name of the cmip6 model.}

\item{.plant_date}{integer, planting date for crop from cell_data_table.}

\item{.harv_date}{integer, harvest data for crop from cell_data_table.}

\item{out.dir}{directory to create in out.path, set with arg[2]}
}
\description{
Used for side effect of writing a weather files based on location.
Location is specified as a raster cell number.
}
\details{
This function creates a weather statistics files for a specified climate model, ssp, and location.
Raw climate model data was converted into daily values in the Gregorian calendar, which sometimes
requires date conversion from the native gcm calendar.  Daily vectors are constructed from 2015 through
to 2100 for each of, minimum daily temperature (tasmin), maximum daily temperature (tasmax),
and precipitation (pr).  These vectors are written as columns in the weather file which is written to
a uniquely named file.  File naming is unique to allow parallel execution of DayCent, with each instance
using a different weather file.

There are 3 files created in this function:
1. Growing season weather statistics using daily means, five-year intervals AND simulation period.
This file includes data on quantiles, means, and standard deviations for tasmin, tasmax, and pr.
Estimates were made over the growing season for the entire simulation period and in five-year intervals.

2. Non-growing season weather statistics using daily means, five-year intervals AND simulation period.
This file includes data on quantiles, means, and standard deviations for tasmin, tasmax, and pr.
Estimates were made over the non-growin season for the entire simulation period and in five-year intervals.

3. Histogram bins of growing and non-growing season variable frequencies, five-year intervals AND simulation period.
This file includes frequency data on tasmin, tasmax, and pr, including consecutive days of pr,
consecutive dry days, consecutive days exceeding 75% quantile and below 25% quantile for tasmin, tasmax.
}
