% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{make_weather_stats}
\alias{make_weather_stats}
\title{Create weather statistics file}
\usage{
make_weather_stats(
  weather_fname,
  gridid,
  ssp,
  gcm,
  plant_day,
  harvest_day,
  out.dir
)
}
\arguments{
\item{weather_fname}{data.table, containing 86 years (2016-2100) of tasmin, tasmax, and pr data for gridcell.
Must be read in to function as a file.path}

\item{out.dir}{directory to create in out.path, set with arg[2]}

\item{.gridid}{integer, cell number (-180 to 180 longitude) from cell_data_table.}

\item{.gcm}{character (length one) giving name of the cmip6 model.}

\item{.ssp}{character (length one) giving name of ssp scenario.}

\item{.plant_date}{integer, planting date for crop from cell_data_table.}

\item{.harv_date}{integer, harvest data for crop from cell_data_table.}
}
\description{
Used for side effect of writing (appending) to a weather statisics file.
}
