% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{pr_sim_run}
\alias{pr_sim_run}
\title{Print row in simulation}
\usage{
pr_sim_run(.row, .gridid, .ssp, .gcm, .crop, .irr, .scenario)
}
\arguments{
\item{.gridid}{integer, gridid associated with row in data.table}

\item{.ssp}{character, ssp of row in data.table}

\item{.gcm}{character, gcm of row in data.table}

\item{.crop}{character, crop of crow in data.table}

\item{.irr}{integer, binary value of irrigation level}

\item{.scenario}{character, scenario of row in data.table}

\item{row}{integer, current row in data.table}
}
\description{
Track simulation row runs with a print statement.
}
