% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{results_processing}
\alias{results_processing}
\title{Prepare DayCent output for pre-processing}
\usage{
results_processing(
  .lis_fname,
  .gridid,
  .scenario,
  .crop,
  .irr,
  .ssp,
  .gcm,
  .plant_date,
  .harv_date,
  .clim_ipcc,
  .argsgcm.ssp,
  .main_out_path,
  .arg_rowstart,
  .arg_endrow,
  .out_tbl_fname,
  .out_gr_tbl_fname
)
}
\arguments{
\item{.lis_fname}{character, lis file name}

\item{.gridid}{integer, gridid associated with row in data.table}

\item{.scenario}{character, scenario associated with row in data.table}

\item{.crop}{character, crop associated with row in data.table}

\item{.irr}{integer, binary irrigation level associated with row in data.table}

\item{.ssp}{character, ssp associated with row in data.table}

\item{.gcm}{character, gcm associated with row in data.table}

\item{.plant_date}{integer, doy of planting associated with row in data.table}

\item{.harv_date}{integer, doy of harvest associated with row in data.table}

\item{.clim_ipcc}{integer, value associated with ipcc climate type from row in data.table}

\item{.argsgcm.ssp}{character, first argument giving name of directory for gcm-ssp}

\item{.main_out_path}{character, directory to write results table to}

\item{.arg_rowstart}{integer, fifth argument giving start row of simulation}

\item{.arg_endrow}{integer, sixth argument giving end row of simulation}

\item{.out_tbl_fname}{character, name of annual results file}

\item{.out_gr_tbl_fname}{character, name of growing, non-growing season results file}
}
\description{
Reformat DayCent output and save results to two tables: annual results, growing season results
}
