% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{revise_eq_sch}
\alias{revise_eq_sch}
\title{Create a revised equilibrium schedule file}
\usage{
revise_eq_sch(schedule.file1_fname, schedule.file2)
}
\arguments{
\item{schedule.file1_fname}{file, extracted from EQ data.table}

\item{schedule.file2}{file, extracted previously from EQ data.table}
}
\description{
Used for revising EQ schedule file for simulation.
}
\details{
This function creates a universal equilibrium schedule file combining both sequences for a grid cell based on a land use ID.
}
