% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{schedule_template}
\alias{schedule_template}
\title{Schedule Template.}
\format{
## `schedule_template`
A data frame with 2812 rows and 8 columns:
\describe{
  \item{scenario}{Scenario in simulation; five levels}
  \item{N_or_S}{North or South Hemisphere - loosely defined.}
  \item{irr}{Irrgiation, binary}
  \item{maiz}{Maize, binary}
  \item{soyb}{Soybean, binary.}
  \item{wwht}{Winter wheat, binary.}
  \item{swht}{Spring wheat, binary.}
  \item{schedule}{Schedule lines relevant for the scenario.}
  ...
}
}
\source{
Cornell University, S.C. McClelland
}
\usage{
schedule_template
}
\description{
Schedule assignments for scenarios in DayCent simulations. Follows sch file format.
Does not include rewild scenario
}
\keyword{datasets}
