% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sql.R
\name{select.tbl}
\alias{select.tbl}
\title{Select a table from MySQL database}
\usage{
select.tbl(con, cohort)
}
\arguments{
\item{con}{connection created to MySQL database}

\item{cohort}{the cohort table (ie 'lu')}
}
\description{
Select a table from MySQL database
}
