% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim_time_logf}
\alias{sim_time_logf}
\title{Track simulation times}
\usage{
sim_time_logf(
  .row,
  .old,
  .end.daycent,
  .start.daycent,
  .gridid,
  .gcm,
  .ssp,
  .crop,
  .irr,
  .scenario,
  .log.file
)
}
\arguments{
\item{.row}{integer,}

\item{.old}{time, start of simulation}

\item{.end.daycent}{time, end of daycent execution}

\item{.start.daycent}{time, start of daycent execution}

\item{.gridid}{integer, gridid associated with row in data.table}

\item{.gcm}{character, gcm associated with row in data.table}

\item{.ssp}{character, ssp associated with row in data.table}

\item{.crop}{character, crop associated with row in data.table}

\item{.irr}{integer, binary irrigation level associated with row in data.table}

\item{.scenario}{character, scenario associated with row in data.table}

\item{.log.file}{character, name of log file to write to}
}
\description{
Write row run time to a log file to track length of each simulation
}
