% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sql.R
\name{sql.query}
\alias{sql.query}
\title{Query table in MySQL database}
\usage{
sql.query(con, table, run_seq_data)
}
\arguments{
\item{con}{connection created to MySQL database}

\item{table}{the cohort table in MySQL database}

\item{run_seq_data}{vector of run_sequences to extract}
}
\description{
Query table in MySQL database
}
