% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{weather_pr_mm_to_cm}
\alias{weather_pr_mm_to_cm}
\title{Fix weather file precipitation}
\usage{
weather_pr_mm_to_cm(.wthfile)
}
\arguments{
\item{.wthfile}{character, name of weather file in working directory}
}
\description{
Used to convert precipitation in weather files (.wth) from mm/day to cm/day.
}
