% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{write_new_crop100_t}
\alias{write_new_crop100_t}
\title{Check crop.100 file}
\usage{
write_new_crop100_t(.crop_100)
}
\arguments{
\item{.gridid}{integer, gridid associated with row in data.table}

\item{.p_gridid}{integer, gridid of previous row}

\item{.crop}{character, crop associated with row in data.table}

\item{.p_crop}{character, crop of previous row}

\item{.simrow}{data.table, single row data.table}

\item{.date}{character, date of test simulation}

\item{.argsgcm.ssp}{character, first argument giving name of directory for gcm-ssp}

\item{.arg_rowstart}{integer, fifth argument giving start row of simulation}

\item{.arg_endrow}{integer, sixth argument giving end row of simulation}
}
\description{
Rewrite the crop.100 if it does not match the gridid and crop in the simulation row
}
