% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{write_new_omad100_t}
\alias{write_new_omad100_t}
\title{Check omad.100 file}
\usage{
write_new_omad100_t(
  .gridid,
  .p_gridid,
  .simrow,
  .date,
  .argsgcm.ssp,
  .arg_rowstart,
  .arg_rowend
)
}
\arguments{
\item{.gridid}{integer, gridid associated with row in data.table}

\item{.p_gridid}{integer, gridid of previous row}

\item{.simrow}{data.table, single row data.table}

\item{.date}{character, date of test simulation}

\item{.argsgcm.ssp}{character, first argument giving name of directory for gcm-ssp}

\item{.arg_rowstart}{integer, fifth argument giving start row of simulation}

\item{.arg_endrow}{integer, sixth argument giving end row of simulation}
}
\description{
Rewrite the omad.100 if it does not match the gridid in the simulation row
}
