% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{write_new_site100}
\alias{write_new_site100}
\title{Check site.100 file}
\usage{
write_new_site100(
  .gridid,
  .p_gridid,
  .crop,
  .p_crop,
  .irr,
  .p_irr,
  .run_seq,
  .argsgcm.ssp,
  .arg_rowstart,
  .arg_rowend,
  sitefile
)
}
\arguments{
\item{.gridid}{integer, gridid associated with row in data.table}

\item{.p_gridid}{integer, gridid of previous row}

\item{.crop}{character, crop associated with row in data.table}

\item{.p_crop}{character, crop of previous row}

\item{.irr}{integer, binary irrigation level associated with row in data.table}

\item{.p_irr}{integer, irr level of previous row}

\item{.run_seq}{integer, run sequence number associated with row in data.table}

\item{.argsgcm.ssp}{character, first argument giving name of directory for gcm-ssp}

\item{.arg_rowstart}{integer, fifth argument giving start row of simulation}

\item{sitefile}{character, current site.100 file}

\item{.arg_endrow}{integer, sixth argument giving end row of simulation}
}
\description{
Rewrite the site.100 if it does not match the gridid, crop, irr in the simulation row
}
