% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{wthfile_exists_wfname}
\alias{wthfile_exists_wfname}
\title{Change weather file name}
\usage{
wthfile_exists_wfname(
  .gridid,
  .ssp,
  .gcm,
  .p_gridid,
  .p_ssp,
  .p_gcm,
  .w_fname,
  arg.gcm,
  arg.ssp
)
}
\arguments{
\item{.gridid}{integer, gridid associated with row in data.table}

\item{.ssp}{character, ssp of row in data.table}

\item{.gcm}{character, gcm of row in data.table}

\item{.p_gridid}{integer, gridid of previous row}

\item{.p_ssp}{character, ssp of previous row}

\item{.p_gcm}{character, gcm of previous row}

\item{.w_fname}{character, name of weather file in tmp directory}

\item{arg.gcm}{character, third argument defined in bash script}

\item{arg.ssp}{character, fourth argument defined in bash script}
}
\description{
Determines if weather file should be removed and new weather file added based on contents
of simulation data.table.
}
