% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{wthfile_exists_wpath}
\alias{wthfile_exists_wpath}
\title{Check if weather file exists in directory}
\usage{
wthfile_exists_wpath(
  .w_fname.path,
  .start_yr,
  .end_yr,
  .gridid,
  .gridid_rotated,
  .ssp,
  .gcm,
  .grab_wth_path,
  .argsgcm.ssp
)
}
\arguments{
\item{.w_fname.path}{character, directory of weather file}

\item{.start_yr}{vector, start year for weather file}

\item{.end_yr}{vector, end year for weather file}

\item{.gridid}{integer, gridid associated with row in data.table}

\item{.gridid_rotated}{integer, rotated gridid associated with row in data.table}

\item{.ssp}{character, ssp of row in data.table}

\item{.gcm}{character, gcm of row in data.table}

\item{.grab_wth_path}{character, directory to write wth file to}

\item{.argsgcm.ssp}{character, second argument defined in bash script}
}
\description{
Checks the weather directory specified for the existence of weather file. If the file
does not exist this function creates the specified weather file and saves it.
}
