% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{create_eq_sch}
\alias{create_eq_sch}
\title{Create an equilibrium schedule file from a template}
\usage{
create_eq_sch(
  cell_data,
  eq_pnh_template = copy(eq_pnh_sch_template),
  eq_psh_template = copy(eq_psh_sch_template),
  .ssp,
  .site_fname,
  .eqtype,
  weather_fname,
  tmp.dir
)
}
\arguments{
\item{cell_data}{multiple row data.table providing input data for the simulation.}

\item{eq_pnh_template}{data.table providing template of schedule file for pnh land use.}

\item{eq_psh_template}{data.table providing template of schedule file for psh land use.}

\item{.ssp}{character, specifies scenario}

\item{.site_fname}{character, name of site.100 file}

\item{.eqtype}{character, either pnh or psh followed by land use (lu) number}

\item{weather_fname}{character, weather file name}

\item{tmp.dir}{character, temporary directory to write the schedule files to}
}
\description{
Used for revising EQ schedule file for simulation.
}
\details{
This function creates a universal equilibrium schedule file for a grid cell based on a land use ID.
}
