% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site.R
\name{create_site}
\alias{create_site}
\title{Create a DayCent site file}
\usage{
create_site(
  .gridid,
  .run_seq,
  .irr,
  cell_data_site = copy(cell_data_site),
  tmp.dir
)
}
\arguments{
\item{.irr}{Integer, irrigation level}

\item{cell}{Integer, grid cell ID number}

\item{cell_data}{Data table, contains spatial data associated with each cell and site.100 file}
}
\description{
Used for writing a site file using spatial data.
}
\details{
This function creates a universal site file for a grid cell.
}
